Updated Android.mk build file
Upon investigation it was discovered that the make file was giving
instructions that it did not need.  If LOCAL_SDK_VERSION,
LOCAL_NDK_VERSION, and LOCAL_NDK_STL_VARIANT are all specified then
there is no need to manually add the STL headers to the list of C
includes.  Unfortunately the same is not true when building an
executable file (i.e. the alljoyn_daemon) so the include headers must
still be specified there.

Also removed where stlport was being included for libAllJoynAndroidExt.so
it was not needed and shouldn't have been included since the code should
be using gnustl if the STL is needed.

Change-Id: I7ba253b724ae48700b6e438b099b96bbe18e0bb9